chore: Bounded Context 분리 기준에 맞게 루트 도메인명 수정 #87
