+++ src/main/java/com/carumuch/capstone/damage/domain/DamageReportRegisteredEvent.java
@@ -1,0 +1,4 @@
+package com.carumuch.capstone.damage.domain;
+
+public record DamageReportRegisteredEvent(DamageReport damageReport) {
+}
