+++ src/main/java/com/carumuch/capstone/damage/domain/DamageReportRepository.java
@@ -1,0 +1,9 @@
+package com.carumuch.capstone.damage.domain;
+
+import java.util.Optional;
+
+public interface DamageReportRepository {
+	DamageReport save(DamageReport damageReport);
+	void deleteById(Long id);
+	Optional<DamageReport> findById(Long id);
+}
