+++ src/main/java/com/carumuch/capstone/damage/presentation/dto/request/UpdateVehicleRequest.java
@@ -1,0 +1,14 @@
+package com.carumuch.capstone.damage.presentation.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record UpdateVehicleRequest(
+	@NotBlank String licenseNumber,
+	@NotBlank String ownershipType,
+	@NotBlank String brand,
+	@NotNull Integer modelYear,
+	@NotBlank String modelName,
+	@NotBlank String ownerName
+) {
+}
