+++ src/main/java/com/carumuch/capstone/identity/domain/user/User.java
@@ -5,7 +5,6 @@ import com.carumuch.capstone.bodyshop.domain.BodyShop;
 import com.carumuch.capstone.community.domain.Comment;
 import com.carumuch.capstone.common.domain.BaseEntity;
 import com.carumuch.capstone.estimate.domain.estimate.Estimate;
-import com.carumuch.capstone.vehicle.domain.Vehicle;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
