--- src/main/java/com/carumuch/capstone/vehicle/domain/DamageReportRegisteredEvent.java
@@ -1,4 +1,0 @@
-package com.carumuch.capstone.vehicle.domain;
-
-public record DamageReportRegisteredEvent(DamageReport damageReport) {
-}
