--- src/main/java/com/carumuch/capstone/vehicle/domain/DamageReportRepository.java
@@ -1,9 +1,0 @@
-package com.carumuch.capstone.vehicle.domain;
-
-import java.util.Optional;
-
-public interface DamageReportRepository {
-	DamageReport save(DamageReport damageReport);
-	void deleteById(Long id);
-	Optional<DamageReport> findById(Long id);
-}
