--- src/main/java/com/carumuch/capstone/vehicle/presentation/dto/request/UpdateVehicleRequest.java
@@ -1,14 +1,0 @@
-package com.carumuch.capstone.vehicle.presentation.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record UpdateVehicleRequest(
-	@NotBlank String licenseNumber,
-	@NotBlank String ownershipType,
-	@NotBlank String brand,
-	@NotNull Integer modelYear,
-	@NotBlank String modelName,
-	@NotBlank String ownerName
-) {
-}
