--- src/main/java/com/carumuch/capstone/vehicle/presentation/dto/response/VehicleInfoResponse.java
@@ -1,25 +1,0 @@
-package com.carumuch.capstone.vehicle.presentation.dto.response;
-
-import com.carumuch.capstone.vehicle.domain.Vehicle;
-
-public record VehicleInfoResponse(
-	Long id,
-	String licenseNumber,
-	String ownershipType,
-	String brand,
-	int modelYear,
-	String modelName,
-	String ownerName
-) {
-	public static VehicleInfoResponse from(Vehicle vehicle) {
-		return new VehicleInfoResponse(
-			vehicle.getId(),
-			vehicle.getLicenseNumber(),
-			vehicle.getOwnershipType().name(),
-			vehicle.getBrand(),
-			vehicle.getModelYear(),
-			vehicle.getModelName(),
-			vehicle.getOwnerName()
-		);
-	}
-}
