+++ src/test/java/com/carumuch/capstone/support/fixture/VehicleFixture.java
@@ -1,7 +1,7 @@
 package com.carumuch.capstone.support.fixture;
 
-import com.carumuch.capstone.vehicle.domain.Vehicle;
-import com.carumuch.capstone.vehicle.domain.VehicleOwnershipType;
+import com.carumuch.capstone.damage.domain.Vehicle;
+import com.carumuch.capstone.damage.domain.VehicleOwnershipType;
 
 public enum VehicleFixture {
 	VEHICLE_FIXTURE_1("11가1111", "기아", 2000, "sm3", "홍길동"),
